refactor(types): modernize types-only package config
- Replace module:commonjs with module:NodeNext (from @tsconfig/node22)
- Add emitDeclarationOnly:true — no JS emitted, only .d.ts
- Remove redundant fields: main, target, lib, esModuleInterop
- Add type:module, exports only exposes types condition
- Collapse 7 single-line enum files into src/index.ts
- Remove src/enums/ subdirectory
