+++ packages/types/package.json
@@ -2,12 +2,11 @@
   "name": "@soker90/finper-types",
   "version": "1.0.0",
   "description": "Shared TypeScript types for finper",
-  "main": "dist/index.js",
+  "type": "module",
   "types": "dist/index.d.ts",
   "exports": {
     ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.js"
+      "types": "./dist/index.d.ts"
     }
   },
   "scripts": {
