--- packages/types/src/enums/budget-rule.ts
@@ -1,1 +1,0 @@
-export type BudgetRuleClassType = 'needs' | 'wants' | 'savings' | 'none'
