--- packages/types/src/enums/debt.ts
@@ -1,1 +1,0 @@
-export type DebtType = 'from' | 'to'
