--- packages/types/src/enums/goal.ts
@@ -1,2 +1,0 @@
-export type GoalColor = '#4CAF50' | '#2196F3' | '#9C27B0' | '#FF9800' | '#F44336' | '#00BCD4' | '#795548' | '#607D8B' | '#E91E63' | '#FFC107'
-export type GoalIcon = 'DollarOutlined' | 'HomeOutlined' | 'CarOutlined' | 'LaptopOutlined' | 'HeartOutlined' | 'RocketOutlined' | 'GiftOutlined' | 'BankOutlined' | 'TrophyOutlined' | 'StarOutlined'
