--- packages/types/src/enums/loan.ts
@@ -1,1 +1,0 @@
-export type LoanPaymentType = 'ordinary' | 'extraordinary'
