--- packages/types/src/enums/stock.ts
@@ -1,1 +1,0 @@
-export type StockType = 'buy' | 'sell' | 'dividend'
