--- packages/types/src/enums/supply.ts
@@ -1,1 +1,0 @@
-export type SupplyType = 'electricity' | 'water' | 'gas' | 'other'
