--- packages/types/src/enums/transaction.ts
@@ -1,1 +1,0 @@
-export type TransactionType = 'expense' | 'income' | 'not_computable'
