+++ packages/types/src/index.ts
@@ -1,7 +1,8 @@
-export type { TransactionType } from './enums/transaction'
-export type { DebtType } from './enums/debt'
-export type { LoanPaymentType } from './enums/loan'
-export type { SupplyType } from './enums/supply'
-export type { StockType } from './enums/stock'
-export type { GoalColor, GoalIcon } from './enums/goal'
-export type { BudgetRuleClassType } from './enums/budget-rule'
+export type TransactionType = 'expense' | 'income' | 'not_computable'
+export type DebtType = 'from' | 'to'
+export type GoalColor = '#4CAF50' | '#2196F3' | '#9C27B0' | '#FF9800' | '#F44336' | '#00BCD4' | '#795548' | '#607D8B' | '#E91E63' | '#FFC107'
+export type GoalIcon = 'DollarOutlined' | 'HomeOutlined' | 'CarOutlined' | 'LaptopOutlined' | 'HeartOutlined' | 'RocketOutlined' | 'GiftOutlined' | 'BankOutlined' | 'TrophyOutlined' | 'StarOutlined'
+export type StockType = 'buy' | 'sell' | 'dividend'
+export type SupplyType = 'electricity' | 'water' | 'gas' | 'other'
+export type LoanPaymentType = 'ordinary' | 'extraordinary'
+export type BudgetRuleClassType = 'needs' | 'wants' | 'savings' | 'none'
